autohighlight: many updates

+++ autohighlight/ahlextract.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import argparse
 import csv
 import json
 import subprocess
@@ -7,7 +8,8 @@ from datetime import datetime, timedelta
 from pathlib import Path
 
 from onglog import find_onglog_entry_by_datetime
-from tdvutil import sec_to_hms, sec_to_shortstr
+from tdvutil import ppretty, sec_to_hms, sec_to_shortstr
+from tdvutil.argparse import CheckFile
 
 # give ourselves a place to stuff our indexes
 script_dir = Path(__file__).parent.resolve()
@@ -15,12 +17,22 @@ INDEX_DIR = script_dir / "indexes"
 INDEX_DIR.mkdir(exist_ok=True)
 
 
+def log(msg):
+    print(msg)
+    sys.stdout.flush()
+
+
 def load_metadata(remux_dir):
     metadata = {}
     for meta_file in Path(remux_dir).glob('*.meta'):
-        with open(meta_file, 'r') as f:
-            data = json.load(f)
+        log(f"INFO: Loading metadata from {meta_file}")
+        try:
+            with open(meta_file, 'r') as f:
+                data = json.load(f)
             metadata[data['filename']] = data
+        except json.JSONDecodeError:
+            log(f"WARNING: Couldn't decode metadata from {meta_file}")
+
     return metadata
 
 
@@ -33,7 +45,7 @@ def load_highlight_requests(index_file):
     return requests
 
 
-def find_highlights(metadata, requests):
+def find_highlights(metadata, requests, content_class):
     highlights = []
     for request in requests:
         if request["completed"] == "Y":
@@ -41,62 +53,183 @@ def find_highlights(metadata, requests):
 
         request_time = datetime.strptime(request['timestamp'], "%Y-%m-%d %H:%M:%S.%f")
         for filename, data in metadata.items():
+            if "content_class" not in data or data['content_class'] != content_class:
+                continue
+            if not data["start_time"]:
+                continue
             start_time = datetime.strptime(data['start_time'], "%Y-%m-%d %H:%M:%S.%f")
             end_time = datetime.strptime(data['end_time'], "%Y-%m-%d %H:%M:%S.%f")
             # print(f"Comparing: {start_time} <= {request_time} <= {end_time}")
             if start_time <= request_time <= end_time:
-                highlights.append((filename, request['timestamp'], request['highlight_id']))
+                highlights.append((filename, request['timestamp'], request['highlight_id'], content_class))
     return highlights
 
 
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Ginerate Ong video metadata",
+    )
+
+    parser.add_argument(
+        "--index",
+        type=Path,
+        default=INDEX_DIR / "ahlindex.csv",
+        action=CheckFile(extensions={"csv"}, must_exist=True),
+        help="The ongoing autohighlight index to read/update"
+    )
+
+    parser.add_argument(
+        "--dest-dir",
+        type=Path,
+        default=Path("."),
+        action=CheckFile(must_exist=True),
+        help="Directory to extract video files into"
+    )
+
+    parser.add_argument(
+        "--source-dir",
+        type=Path,
+        default=Path("."),
+        action=CheckFile(must_exist=True),
+        help="Directory to look for source videos to extract"
+    )
+
+    parser.add_argument(
+        "--content-class",
+        type=str,
+        default="clean",
+        help="Content class to extract"
+    )
+
+    parser.add_argument(
+        "--extra-head",
+        type=int,
+        default=3,
+        help="Extra seconds to add to the start of the highlight"
+    )
+
+    parser.add_argument(
+        "--extract-length", "--extract-len",
+        type=int,
+        default=60,
+        help="Number of seconds to extract as highlight"
+    )
+
+    # parser.add_argument(
+    #     "--trash-dir",
+    #     type=Path,
+    #     default=None,
+    #     action=CheckFile(must_exist=True),
+    #     help="Directory to move video files into when done",
+    # )
+
+    parser.add_argument(
+        "--force",
+        action="store_true",
+        default=False,
+        help="Overwrite existing output files",
+    )
+
+    parser.add_argument(
+        "--force-metadata",
+        action="store_true",
+        default=False,
+        help="Overwrite existing metadata files",
+    )
+
+    # parser.add_argument(
+    #     "--force-remux",
+    #     action="store_true",
+    #     default=False,
+    #     help="Overwrite existing remuxed video",
+    # )
+
+    # parser.add_argument(
+    #     "--keep-original",
+    #     action="store_true",
+    #     default=True,
+    #     help="Keep original video file",
+    # )
+
+    # parser.add_argument(
+    #     "filenames",
+    #     type=Path,
+    #     nargs="+",
+    #     metavar="filename",
+    #     # action=CheckFile(extensions=valid_extensions, must_exist=True),
+    #     help="video file(s) to process",
+    # )
+
+    parser.add_argument(
+        "--debug",
+        action="store_true",
+        default=False,
+        help="Enable debug output",
+    )
+
+    parser.add_argument(
+        "--dry-run",
+        action="store_true",
+        default=False,
+        help="Dry run, don't actually remux anything",
+    )
+
+    parsed_args = parser.parse_args()
+
+    return parsed_args
+
+
 def main():
-    remux_dir = Path("x:/zTEMP/remux_tmp")
-    index_file = INDEX_DIR / "ahlindex.csv"
-    output_index_file = INDEX_DIR / "output_index.csv"
+    args = parse_args()
     autohighlight_dir = Path("x:/zTEMP/autohighlight_tmp")
 
-    metadata = load_metadata(remux_dir)
-    requests = load_highlight_requests(index_file)
+    metadata = load_metadata(args.source_dir)
+    requests = load_highlight_requests(args.index)
 
-    print("metadata: " + str(metadata))
-    print("\nrequests: " + str(requests))
-    print()
-    highlights = find_highlights(metadata, requests)
+    if args.debug:
+        log(f"DEBUG: metadata: {ppretty(metadata)}\n")
+        log(f"DEBUG: requests: {ppretty(requests)}\n")
 
-    print("Videos containing requested highlights:")
-    for filename, timestamp, description in highlights:
-        print(f"- {filename}: {timestamp} - {description}")
+    highlights = find_highlights(metadata, requests, args.content_class)
 
-    output_meta = {}
-    for filename, timestamp, highlight_id in highlights:
-        video_path = remux_dir / filename
+    log(f"INFO: Found {len(highlights)} highlight segments out of {len(requests)} requests")
+
+    if args.debug:
+        log("DEBUG: Videos containing requested highlights:")
+        for filename, timestamp, description, content_class in highlights:
+            log(f"  - {filename}: {timestamp} - {description}")
+
+    for filename, timestamp, highlight_id, content_class in highlights:
+        output_meta = {}
+        video_path = args.source_dir / filename
         request_time = datetime.strptime(timestamp, "%Y-%m-%d %H:%M:%S.%f")
         start_time = datetime.strptime(metadata[filename]['start_time'], "%Y-%m-%d %H:%M:%S.%f")
         start_date = metadata[filename]['start_time'].split(" ")[0]
 
-        time_offset = (request_time - start_time).total_seconds() + 3
-        offset_str = sec_to_hms(time_offset).split(".")[0]
+        time_offset = (request_time - start_time).total_seconds() + 3 - args.extra_head
+        # offset_str = sec_to_hms(time_offset).split(".")[0]
 
-        output_filename = f"highlight_{highlight_id}_{start_date}_uptime_{sec_to_shortstr(time_offset)}.mp4"
+        input_ext = Path(filename).suffix
+
+        output_filename = f"highlight_{highlight_id}_{content_class}_{start_date}{input_ext}"
         output_path = autohighlight_dir / output_filename
 
-        metadata_filename = f"highlight_{highlight_id}_{start_date}_uptime_{sec_to_shortstr(time_offset)}.txt"
+        metadata_filename = f"highlight_{highlight_id}_meta_{start_date}.txt"
         metadata_path = autohighlight_dir / metadata_filename
 
-
-
-        if output_path.exists():
-            print(f"Highlight {highlight_id} from {filename} already exists. Skipping.")
+        if output_path.exists() and not args.force:
+            log(f"INFO: Highlight {highlight_id} from {filename} already exists. Skipping.")
             continue
 
         ffmpeg_cmd = [
             "ffmpeg",
             "-ss", f"{time_offset}",
-            "-t", "60",
+            "-t", str(args.extra_head + args.extract_length),
             "-i", str(video_path),
             "-vsync", "vfr",
             "-c:v", "h264_nvenc",
             "-preset", "p3",
+            "-rc", "constqp",
             "-qp", "16",
             "-b:v", "0",
             "-c:a", "libfdk_aac", "-vbr", "5", "-cutoff", "18000",
@@ -105,32 +238,43 @@ def main():
             str(output_path)
         ]
 
-        print(f"Extracting highlight {highlight_id} from {filename} at offset {time_offset}")
-        sys.stdout.flush()
-        # print(f"Command: {' '.join(ffmpeg_cmd)}")
+        log(f"INFO: Extracting highlight {highlight_id} from {filename} at offset {time_offset}")
+        if args.debug:
+            log(f"DEBUG: Command: {' '.join(ffmpeg_cmd)}")
+
+        if args.dry_run:
+            log(f"DRY RUN: Would have extracted highlight to {output_path}")
+            continue
 
         try:
             subprocess.run(ffmpeg_cmd, check=True, capture_output=True, text=True)
-            print(f"Successfully extracted highlight to {output_filename}")
+            log(f"INFO: Successfully extracted highlight to {output_filename}")
         except subprocess.CalledProcessError as e:
-            print(f"Error extracting highlight: {e}")
-            print(f"ffmpeg stderr: {e.stderr}")
+            log(f"ERROR: Couldn't extract highlight: {e}")
+            log(f"ERROR: ffmpeg stderr: {e.stderr}")
             continue
 
         output_meta["highlight_id"] = highlight_id
         output_meta["request_time"] = str(request_time)
         output_meta["request_uptime"] = sec_to_hms(time_offset)
         output_meta["original_filename"] = filename
-
+        output_meta["extra_head"] = args.extra_head
+        output_meta["extract_length"] = args.extract_length
         logentry = find_onglog_entry_by_datetime(request_time)
-        # print(f"logentry: {logentry}")
+
         if logentry:
+            if args.debug:
+                log(f"DEBUG: found onglog entry: {ppretty(logentry)}")
             output_meta["title"] = logentry.title
             output_meta["requester"] = logentry.requester
+            output_meta["onglog_line"] = logentry.onglog_line_number
+        else:
+            log(f"WARNING: Couldn't find onglog entry for {request_time}")
 
-        metadata_path.write_text(json.dumps(output_meta, indent=4))
+        if not metadata_path.exists() or args.force_metadata:
+            metadata_path.write_text(json.dumps(output_meta, indent=4))
 
-    print("\nHighlight extraction complete.")
+    log("\nINFO: Highlight extraction complete.")
     # output_index_file.write_text("\n".join(meta_lines))
 
 if __name__ == "__main__":
